chore(payments-server): move dialog message component into directory
- refs #3218
