+++ packages/fxa-payments-server/src/components/DialogMessage/index.scss
@@ -1,5 +1,5 @@
-@import '../../../fxa-content-server/app/styles/breakpoints';
-@import '../../../fxa-content-server/app/styles/variables';
+@import '../../../../fxa-content-server/app/styles/breakpoints';
+@import '../../../../fxa-content-server/app/styles/variables';
 
 #dialogs {
   .modal {
